Hfed prepare for reset migrations

+++ MVC5-Seneca/Migrations/201905221555383_AddVolunteerHours.cs
@@ -1,13 +1,12 @@
 namespace MVC5_Seneca.Migrations
 {
-    using System;
     using System.Data.Entity.Migrations;
     
     public partial class AddVolunteerHours : DbMigration
     {
         public override void Up()
         {
-            AddColumn("dbo.HfedSchedule", "VolunteerHours", c => c.Int());
+            AddColumn("dbo.HfedSchedule", "VolunteerHours", c => c.Single());
         }
         
         public override void Down()
